ctf-writer: Split clock and clock-class declarations and definitions
Split the clock_class functions of ctf-writer in its own header file.
clock_class deserves it.
