+++ include/babeltrace/ctf-writer/clock.h
@@ -280,59 +280,6 @@ void bt_ctf_clock_put(struct bt_ctf_clock *clock)
 {
         bt_ctf_object_put_ref(clock);
 }
-
-extern struct bt_ctf_clock_class *bt_ctf_clock_class_create(const char *name,
-                uint64_t freq);
-
-extern const char *bt_ctf_clock_class_get_name(
-                struct bt_ctf_clock_class *clock_class);
-
-extern int bt_ctf_clock_class_set_name(struct bt_ctf_clock_class *clock_class,
-                const char *name);
-
-extern const char *bt_ctf_clock_class_get_description(
-                struct bt_ctf_clock_class *clock_class);
-
-extern int bt_ctf_clock_class_set_description(
-                struct bt_ctf_clock_class *clock_class,
-                const char *desc);
-
-extern uint64_t bt_ctf_clock_class_get_frequency(
-                struct bt_ctf_clock_class *clock_class);
-
-extern int bt_ctf_clock_class_set_frequency(
-                struct bt_ctf_clock_class *clock_class, uint64_t freq);
-
-extern uint64_t bt_ctf_clock_class_get_precision(
-                struct bt_ctf_clock_class *clock_class);
-
-extern int bt_ctf_clock_class_set_precision(
-                struct bt_ctf_clock_class *clock_class, uint64_t precision);
-
-extern int bt_ctf_clock_class_get_offset_s(
-                struct bt_ctf_clock_class *clock_class, int64_t *seconds);
-
-extern int bt_ctf_clock_class_set_offset_s(
-                struct bt_ctf_clock_class *clock_class, int64_t seconds);
-
-extern int bt_ctf_clock_class_get_offset_cycles(
-                struct bt_ctf_clock_class *clock_class, int64_t *cycles);
-
-extern int bt_ctf_clock_class_set_offset_cycles(
-                struct bt_ctf_clock_class *clock_class, int64_t cycles);
-
-extern bt_bool bt_ctf_clock_class_is_absolute(
-                struct bt_ctf_clock_class *clock_class);
-
-extern int bt_ctf_clock_class_set_is_absolute(
-                struct bt_ctf_clock_class *clock_class, bt_bool is_absolute);
-
-extern const unsigned char *bt_ctf_clock_class_get_uuid(
-                struct bt_ctf_clock_class *clock_class);
-
-extern int bt_ctf_clock_class_set_uuid(struct bt_ctf_clock_class *clock_class,
-                const unsigned char *uuid);
-
 #ifdef __cplusplus
 }
 #endif
